♻  Rewrote all cmdlets to be "binary cmdlets" (Unreleased at this time)
  This entire commit will result in a breaking change once released.

  While this might be considered unnecessary by some, it did result in
  less memory usage, better safety when modifying the API itself, as
  well as paving the way to better understanding using .NET Core with
  powershell. This also means we can be one step closer to 'binding' C
  and C++ libraries to powershell. This means that nearly anything will
  be bindable to powershell, and that's truly terrifying. :)

TODO: Documentation and descriptions are needed for all cmdlets. This
will be handled soon.

A better README file is needed as well. Currently, there is no Pester
adapter for VSTest and therefore there is no adapter for using `dotnet
test`. This, to put it mildly, is terrible.

I do wish I had the time to investigate this more, but I'll be setting
up a github action system so that git tags will be used to deploy
packages. More work is needed. More work is always needed. Always.